# docs/api/api_v3/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Koudaisai Portal API v3",
    "version": "0.1.0"
  },
  "paths": {
    "/approval-requests": {
      "get": {
        "tags": [
          "approval-requests"
        ],
        "description": "Get all approval requests, optionally filtered by group.",
        "operationId": "get_approval_requests",
        "parameters": [
          {
            "name": "group_id",
            "in": "query",
            "description": "グループ絞り込み(現状はクエリ受理のみ。範囲は呼び出し元の所属グループに限定)。",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApprovalRequestRead"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "approval-requests"
        ],
        "description": "Create an approval request.",
        "operationId": "post_approval_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Invalid approval request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/approval-requests/{id}": {
      "get": {
        "tags": [
          "approval-requests"
        ],
        "description": "Get an approval request by id.",
        "operationId": "get_approval_request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approval request found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Approval request not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "approval-requests"
        ],
        "description": "Delete an approval request by id.",
        "operationId": "delete_approval_request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Approval request deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Approval request not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/approval-requests/{id}/approve": {
      "post": {
        "tags": [
          "approval-requests"
        ],
        "description": "Approve an approval request by id.",
        "operationId": "approve_approval_request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalActionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Approval request approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Approval request not found"
          },
          "409": {
            "description": "Approval request is not pending"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/approval-requests/{id}/close": {
      "post": {
        "tags": [
          "approval-requests"
        ],
        "description": "Close an approval request by id.",
        "operationId": "close_approval_request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approval request closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Approval request not found"
          },
          "409": {
            "description": "Approval request is not pending"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/approval-requests/{id}/reject": {
      "post": {
        "tags": [
          "approval-requests"
        ],
        "description": "Reject an approval request by id.",
        "operationId": "reject_approval_request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalActionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Approval request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequestRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Approval request not found"
          },
          "409": {
            "description": "Approval request is not pending"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovalActionBody": {
        "type": "object",
        "properties": {
          "reason": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ApprovalRequestCreate": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApprovalRequestType"
          },
          {
            "type": "object",
            "required": [
              "issue_reason"
            ],
            "properties": {
              "issue_reason": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ApprovalRequestRead": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApprovalRequestType"
          },
          {
            "$ref": "#/components/schemas/ApprovalRequestStatus"
          },
          {
            "type": "object",
            "required": [
              "id",
              "issued_at",
              "issued_by",
              "issue_reason"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "issue_reason": {
                "type": "string"
              },
              "issued_at": {
                "type": "string",
                "format": "date-time"
              },
              "issued_by": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ]
      },
      "ApprovalRequestStatus": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "status"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "pending"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "approved_by",
              "approved_at",
              "status"
            ],
            "properties": {
              "approval_reason": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "approved_at": {
                "type": "string",
                "format": "date-time"
              },
              "approved_by": {
                "type": "string",
                "format": "uuid"
              },
              "status": {
                "type": "string",
                "enum": [
                  "approved"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "rejected_by",
              "rejected_at",
              "status"
            ],
            "properties": {
              "rejected_at": {
                "type": "string",
                "format": "date-time"
              },
              "rejected_by": {
                "type": "string",
                "format": "uuid"
              },
              "rejection_reason": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "rejected"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "closed_at",
              "status"
            ],
            "properties": {
              "closed_at": {
                "type": "string",
                "format": "date-time"
              },
              "status": {
                "type": "string",
                "enum": [
                  "closed"
                ]
              }
            }
          }
        ]
      },
      "ApprovalRequestType": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "description": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "icon_key": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "type": {
                "type": "string",
                "enum": [
                  "edit_exhibition_info"
                ]
              }
            }
          }
        ]
      },
...
    }
  }
}